Changed error handling to ensure Error object is returned
I want to make sure that the user always receives an Error object from an error. So far I've been channeling all errors through the _handleError function (which also helps to ensure that we check for an onerror function existing).  I've modified that function to check of the first parameter (err) is an error object. If it is, that object is returned. This lets the user receive the actual error, that could possibly come from the IndexedDB api or elsewhere.  If not I'm wrapping the error message in an Error object and returning that.  In the future there might be a customer Error type for gazel, but for now it seems unnecessary.